scripts/requirements.txt
absl-py==1.4.0
ai2thor==5.0.0
ai2thor-colab==0.1.2
aiohttp==3.8.1
aiosignal==1.2.0
allennlp==2.9.0
allennlp-models==2.9.0
aniso8601==9.0.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
argtyped==0.3.1
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
aws-requests-auth==0.4.3
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
backports.csv==1.0.7
base58==2.1.1
beautifulsoup4==4.11.1
bleach==5.0.0
blessed==1.20.0
blinker==1.6.2
blis==0.4.1
boto3==1.21.37
botocore==1.24.37
Brotli==1.0.9
brotlipy==0.7.0
cached-path==1.1.2
cachetools==5.0.0
canonicaljson==2.0.0
catalogue==1.0.0
CD==1.1.0
certifi==2021.10.8
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
cftime==1.6.2
chardet==4.0.0
charset-normalizer==2.0.12
checklist==0.0.11
cheroot==8.6.0
CherryPy==18.6.1
click==8.1.3
clip @ git+https://github.com/openai/CLIP.git@a1d071733d7111c9c014f024669f959182114e33
cloudpickle==2.2.1
colorama @ file:///tmp/build/80754af9/colorama_1607707115595/work
conllu==4.4.1
cryptography==36.0.2
cycler==0.11.0
cymem==2.0.6
Cython==0.29.28
dask==2023.4.1
datasets==2.0.0
debugpy @ file:///tmp/build/80754af9/debugpy_1637091796427/work
decorator==4.4.2
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.4
docker-pycreds==0.4.0
easydict==1.9
ego4d==1.3.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.2.5/en_core_web_sm-2.2.5.tar.gz
entrypoints==0.3
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fairscale==0.4.5
fast-rcnn==0.0.0
fastjsonschema==2.15.3
feather==0.0.8
feather-format==0.4.1
feedparser==6.0.8
filelock==3.4.2
Flask==2.2.2
Flask-Cors==1.1.2
Flask-RESTful==0.3.9
fonttools==4.31.2
frozenlist==1.3.0
fsspec==2022.3.0
ftfy==6.1.1
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
google-api-core==2.7.1
google-auth==2.6.3
google-cloud-core==2.2.3
google-cloud-storage==2.2.1
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.0
gpustat==1.0.0
h11==0.13.0
h5py==3.6.0
huggingface-hub @ file:///tmp/build/80754af9/huggingface_hub_1639662742275/work
idna==2.10
imageio==2.16.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.6.0
importlib-resources==5.6.0
iniconfig==1.1.1
ipykernel @ file:///tmp/build/80754af9/ipykernel_1646982648551/work/dist/ipykernel-6.9.1-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1648799197050/work
ipython-genutils==0.2.0
ipywidgets==7.7.0
iso-639==0.4.5
itsdangerous==2.1.2
jaraco.classes==3.2.1
jaraco.collections==3.5.1
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.7.0
jedi @ file:///tmp/build/80754af9/jedi_1644315233700/work
Jinja2==3.1.2
jmespath==1.0.0
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
jsonnet==0.18.0
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client @ file:///opt/conda/conda-bld/jupyter_client_1643638337975/work
jupyter-console==6.4.3
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1646994441805/work
jupyterlab-pygments==0.2.0
jupyterlab-widgets==1.1.0
kiwisolver==1.4.0
langcodes==3.3.0
lmdb==1.4.0
locket==1.0.0
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline @ file:///tmp/build/80754af9/matplotlib-inline_1628242447089/work
mistune==0.8.4
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
more-itertools==8.12.0
moviepy==1.0.3
msgpack==1.0.4
msgpack-numpy==0.4.8
multidict==6.0.2
multiprocess==0.70.12.2
munch==2.5.0
murmurhash==1.0.6
mutagen==1.46.0
mypy-extensions==0.4.3
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.3.0
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
netCDF4==1.6.3
networkx==2.7.1
nltk==3.6.5
notebook==6.4.10
numpy==1.20.2
nvidia-ml-py==11.495.46
opencv-python==4.5.5.64
outcome==1.1.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.3
pandocfilters==1.5.0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
partd==1.4.0
path==16.6.0
pathtools==0.1.2
pathy==0.6.1
patternfork-nosql==3.6
pdfminer.six==20220319
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.2.0
plac==1.1.3
pluggy==1.0.0
portalocker==2.7.0
portend==3.1.0
preshed==3.0.6
prior==1.0.3
proglog==0.1.10
progressbar2==4.2.0
prometheus-client==0.14.1
promise==2.3
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.19.4
psutil==5.9.0
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
py==1.11.0
py-rouge==1.1
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pycryptodomex==3.18.0
pydantic==1.8.2
PyGithub==1.58.2
Pygments @ file:///opt/conda/conda-bld/pygments_1644249106324/work
PyJWT==2.7.0
PyNaCl==1.5.0
pynvml==11.5.0
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==3.0.7
pyrsistent==0.18.1
pysat==3.0.6
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pytest==7.1.1
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-docx==0.8.11
python-fcl==0.7.0.4
python-sat==0.1.8.dev3
python-slugify==5.0.2
python-utils==3.5.2
python-xlib==0.33
pytube==15.0.0
pytz==2022.1
PyWavelets==1.3.0
PyYAML==5.1
pyzmq @ file:///tmp/build/80754af9/pyzmq_1638436375034/work
qtconsole==5.3.0
QtPy==2.0.1
regex @ file:///tmp/build/80754af9/regex_1648447708112/work
requests==2.25.1
responses==0.18.0
rsa==4.8
s3transfer==0.5.2
sacremoses @ file:///tmp/build/80754af9/sacremoses_1633107328213/work
scikit-image==0.19.2
scikit-learn==1.0.2
scipy==1.8.0
selenium==4.1.3
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.5.8
setproctitle==1.2.2
sgmllib3k==1.0.0
shapely==2.0.1
shortuuid==1.0.8
six @ file:///tmp/build/80754af9/six_1644875935023/work
sklearn==0.0
smart-open==5.2.1
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2
spacy==2.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==1.0.5
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
tempora==5.0.1
tensorboard-data-server==0.6.1
tensorboardX==2.5
termcolor==1.1.0
terminado==0.13.3
testpath==0.6.0
text-unidecode==1.3
thinc==7.4.0
threadpoolctl==3.1.0
tifffile==2022.3.25
timer==0.2.2
tokenizers @ file:///tmp/build/80754af9/tokenizers_1639607818835/work
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
torch==1.10.2
torch-tb-profiler==0.4.0
torchaudio==0.7.0a0+a853dff
torchvision==0.11.3
tornado @ file:///tmp/build/80754af9/tornado_1606942300299/work
tqdm @ file:///opt/conda/conda-bld/tqdm_1647339053476/work
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
transformers==4.5.1
trimesh==3.21.6
trio==0.20.0
trio-websocket==0.9.2
typer==0.4.1
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3==1.26.9
wandb==0.12.14
wasabi==0.9.1
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websockets==11.0.3
Werkzeug==2.3.4
widgetsnbextension==3.6.0
word2number==1.1
wrapt==1.15.0
wsproto==1.1.0
xarray==2023.1.0
xxhash==3.0.0
yarl==1.7.2
yt-dlp==2023.7.6
zc.lockfile==2.0
zipp @ file:///opt/conda/conda-bld/zipp_1641824620731/work
